fix dependency chain issues
